refactor(redskilled): isolate ACP GitHub projection Refs #3886
